Renamed Test files to mirror DAO tests.
